// PAKNWEB/src/app/models/SystemConfigObject.ts
export class SystemConfigObject {
  //constructor() {
  //  this.Id = -1;
  //  this.code = "";
  //  this.trangThai = true;
  //  this.tenCauHinh = "";
  //  this.giaTri = "";
  //  this.moTa = "";
  //  this.maDonVi = -1;
  //}
  Id: number = 0;
  code: string = '';
  trangThai: boolean = true;
  tenCauHinh: string = '';
  giaTri: string = '';
  moTa: string = '';
  maDonVi: number = -1;

}